feat: show author for posts that specify one
Posts by default inherit the site's author, but it's possible to
override that per post, which affects--e.g.--the RSS metadata. For posts
that do so, we don't currently show that anywhere in the theme.

Add a new piece of metadata to posts, along with the date and reading
time, that displays the author only when it's not inherited from the
site.