Revert "fix: document does not have a valid `rel=canonical`"
The SEO plugin adds a rel="canonical", so when it's enabled this makes
us emit multiple, which isn't allowed. When the SEO plugin isn't
enabled, it's fine not to emit a rel="canonical", since that tag exists
only for SEO.

Additionally, the value of this canonical tag is wrong. It always points
to the root of the site, but it's supposed to point to the main URL of
the current page.

This reverts commit f75cca29781611428a8e8a8e2ac9b170365120be.